Add @grails.validation.Validateable to All command objects and put them in a new class to avoid issue with inner classes.
