+++ grails-app/controllers/grails/plugin/springsecurity/ui/RegisterController.groovy
@@ -246,45 +246,3 @@ class RegisterController extends AbstractS2UiController {
 		}
 	}
 }
-
-class ForgotPasswordCommand implements CommandObject {
-	String username
-}
-
-class RegisterCommand implements CommandObject {
-
-	protected static Class<?> User
-	protected static String usernamePropertyName
-
-	String username
-	String email
-	String password
-	String password2
-
-	static constraints = {
-		username validator: { value, command ->
-			if (!value) {
-				return
-			}
-
-			if (User.findWhere((usernamePropertyName): value)) {
-				return 'registerCommand.username.unique'
-			}
-		}
-		email email: true
-		password validator: RegisterController.passwordValidator
-		password2 nullable: true, validator: RegisterController.password2Validator
-	}
-}
-
-class ResetPasswordCommand implements CommandObject {
-
-	String username
-	String password
-	String password2
-
-	static constraints = {
-		password validator: RegisterController.passwordValidator
-		password2 nullable: true, validator: RegisterController.password2Validator
-	}
-}
