+++ src/groovy/grails/plugin/springsecurity/ui/ForgotPasswordCommand.groovy
@@ -1,0 +1,6 @@
+package grails.plugin.springsecurity.ui
+
+@grails.validation.Validateable
+class ForgotPasswordCommand implements CommandObject {
+	String username
+}
